+++ src/components/DarApplication.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { div, span, hh } from "react-hyperscript-helpers";
-import { Theme } from '../theme';
-import { AppSummary } from '../components/AppSummary';
+import { Theme } from '../libs/theme';
+import { AppSummary } from './AppSummary';
 
 const SECTION = {
   fontFamily: 'Montserrat',
@@ -19,7 +19,7 @@ const HEADER_BOLD = {
   fontWeight: Theme.font.weight.semibold,
 }
 
-export const Application = hh(class Application extends React.PureComponent {
+export const DarApplication = hh(class DarApplication extends React.PureComponent {
   render() {
     const { voteAsChair, darInfo } = this.props;
     return div([
